clearing ESMA from all MITM ralated code
using ramanujan package's code when possible
