+++ ESMA/enumerate_over_signed_rcf.py
@@ -5,10 +5,11 @@ import itertools
 import mpmath
 import sympy
 from sympy import lambdify, Rational
-from series_generators import create_series_from_shift_reg
+from ramanujan.utils.series_generators import create_series_from_shift_reg
 from massey import slow_massey
-from mobius import GeneralizedContinuedFraction, EfficientGCF
-from convergence_rate import calculate_convergence
+from EfficientGCF import EfficientGCF
+from ramanujan.utils.mobius import GeneralizedContinuedFraction
+from ramanujan.utils.convergence_rate import calculate_convergence
 
 """
 Some important terminology:
